+++ database/migrations/2021_01_17_124522_create_obtain_skills_table.php
@@ -15,7 +15,7 @@ class CreateObtainSkillsTable extends Migration
     {
         Schema::create('obtain_skills', function (Blueprint $table) {
             $table->id();
-            $table->integer('course_id')->unsigned()->nullable();
+            $table->bigInteger('course_id')->unsigned()->nullable();
             $table->string('name');
             $table->timestamps();
         });
