+++ database/migrations/2021_01_17_125435_create_gained_experiences_table.php
@@ -15,7 +15,7 @@ class CreateGainedExperiencesTable extends Migration
     {
         Schema::create('gained_experiences', function (Blueprint $table) {
             $table->id();
-            $table->integer('course_id')->unsigned()->nullable();
+            $table->bigInteger('course_id')->unsigned()->nullable();
             $table->string('name');
             $table->timestamps();
         });
