+++ database/migrations/2021_01_17_130737_create_course_lessons_table.php
@@ -15,7 +15,7 @@ class CreateCourseLessonsTable extends Migration
     {
         Schema::create('course_lessons', function (Blueprint $table) {
             $table->id();
-            $table->integer('course_id')->unsigned()->nullable();
+            $table->bigInteger('course_id')->unsigned()->nullable();
             $table->string('title', 100);
             $table->text('description');
             $table->timestamps();
